Fixed JavaScript load order.
The load order for the JavaScript needs to be changed so that the HTML
element that the script is referencing is loaded into the DOM before the
script trys to access it. By moving the script tags to the bottom of the
HTML they are loaded into the DOM after the HTML elements they are
referencing.