Added addItem
Created a method for the add item component to be called by the enter key up event or when the add button is pressed. This fuction takes the text from the component input and adds it to the item list as an item by emitting the add event. The add event is set up as an event listener on the add item component as an attribute which calls the global add item fuction that pushes the text into the items array which updates the list.