Added Hello Component
The hello component is a vue file which contains all of the code for the component. The component is then added to the index html using its name hello-component tag. The component also needs to be loaded in the main js file. The componenet is imported in the main js file and listed in the components for the app.