Added config
The binary state that toggles between working and resting periods of time. This state cannot be toggled by buttons; it should somehow be managed by our application's internal logic. For the sake of modularity and maintainability, let's define the amount of time for work and for rest in an external file. By these initializations I mean that our Pomodoro should count down 20 minutes for the working Pomodoro interval and 5 minutes for breaks.